backend/requirements.txt
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.2
psycopg2-binary==2.9.10
PyJWT==2.11.0
scikit-learn==1.6.1
whitenoise==6.9.0
